feat: implement role management service with user-role assignment logic

+++ src/main/java/com/aseem/lakesidehotel/service/UserService.java
@@ -46,7 +46,10 @@ public class UserService implements IUserService {
     @Transactional
     @Override
     public void deleteUser(String email) {
-        userRepository.deleteByEmail(email);
+        User theUser = getUser(email);
+        if(theUser != null){
+            userRepository.deleteByEmail(email);
+        }
     }
 
     @Override
